Advanced Task

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,10 +8,10 @@
 
 int main(void)
 {
-	long i = 2;
-	long j = 1;
+	unsigned int i = 2;
+	unsigned int j = 1;
 	int _;
-	int temp;
+	unsigned int temp;
 
 	printf("1, 2, ");
 	for (_ = 0; _ < 48; _++)
